@@ -1,15 +1,16 @@
 name: flutter_auto_form
 description: A package simplifying the process of creating user friendly forms in Flutter
-version: 0.2.3
+version: 1.0.0-dev.1
 homepage: https://github.com/GaspardMerten/flutter_auto_form
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.15.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
-  smarter_text_field: ^0.0.4
-  dropdown_search: ^1.0.0
+  smarter_text_field: ^0.0.7
+
+  dropdown_search: ^2.0.1
 
   flutter:
     sdk: flutter
