@@ -1,49 +1,88 @@
 import 'package:dropdown_search/dropdown_search.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_auto_form/flutter_auto_form.dart';
+import 'package:flutter_auto_form/src/models/field/field_context.dart';
+import 'package:flutter_auto_form/src/widgets/fields/interface.dart';
 
-class SearchModelField<T extends Object> extends StatelessWidget {
-  const SearchModelField({
+class SearchModelFieldWidget<T extends Object> extends FieldStatefulWidget {
+  const SearchModelFieldWidget({
     Key? key,
-    required this.label,
-    required this.search,
-    required this.onSelected,
-    this.validator,
-    required this.selectedValue,
-    this.autoValidateMode = AutovalidateMode.onUserInteraction,
+    required this.fieldContext,
   }) : super(key: key);
 
-  final String label;
+  @override
+  final FieldContext fieldContext;
 
-  final Future<List<T>> Function(String? query) search;
+  @override
+  State<SearchModelFieldWidget<T>> createState() =>
+      _SearchModelFieldWidgetState<T>();
+}
 
-  final Function(T? value) onSelected;
-
-  final FormFieldValidator? validator;
-
-  final T? selectedValue;
-
-  final AutovalidateMode autoValidateMode;
+class _SearchModelFieldWidgetState<T extends Object>
+    extends FieldState<SearchModelFieldWidget<T>> {
+  late final AFSearchModelField<T> field =
+      widget.fieldContext.field as AFSearchModelField<T>;
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(
-      height: 50,
-      child: DropdownSearch<T>(
-        label: label,
-        onChanged: onSelected,
-        selectedItem: selectedValue,
-        validator: validator,
-        autoValidateMode: autoValidateMode,
-        onFind: search,
-        mode: Mode.MENU,
-        clearButton: const SizedBox(),
-        isFilteredOnline: true,
-        showClearButton: true,
-        showSearchBox: true,
-        dropdownSearchDecoration: const InputDecoration().applyDefaults(
-          Theme.of(context).inputDecorationTheme,
-        ),
-      ),
+    final decorationTheme = Theme.of(context).inputDecorationTheme;
+
+    return DropdownSearch<T>(
+      selectedItem: field.value,
+      onChanged: onChanged,
+      validator: field.validator,
+      autoValidateMode: widget.fieldContext.forceErrorDisplay
+          ? AutovalidateMode.always
+          : AutovalidateMode.onUserInteraction,
+      onFind: field.search,
+      mode: Mode.MENU,
+      clearButton: const SizedBox(),
+      isFilteredOnline: true,
+      showClearButton: true,
+      showSearchBox: true,
+      dropdownSearchDecoration: InputDecoration(
+        label: Text(widget.fieldContext.field.name),
+      ).applyDefaults(decorationTheme),
+    );
+  }
+}
+
+class SearchMultipleModelsField<T extends Object> extends FieldStatefulWidget {
+  const SearchMultipleModelsField({
+    Key? key,
+    required this.fieldContext,
+  }) : super(key: key);
+
+  @override
+  final FieldContext fieldContext;
+
+  @override
+  State<SearchMultipleModelsField<T>> createState() =>
+      _SearchMultipleModelsFieldState<T>();
+}
+
+class _SearchMultipleModelsFieldState<T extends Object>
+    extends FieldState<SearchMultipleModelsField<T>> {
+  late final AFSearchMultipleModelsField<T> field =
+      widget.fieldContext.field as AFSearchMultipleModelsField<T>;
+
+  @override
+  Widget build(BuildContext context) {
+    final decorationTheme = Theme.of(context).inputDecorationTheme;
+
+    return DropdownSearch.multiSelection(
+      selectedItems: field.value ?? <T>[],
+      onChanged: onChanged,
+      validator: widget.fieldContext.forceErrorDisplay ? field.validator : null,
+      onFind: field.search,
+      mode: Mode.MENU,
+      clearButton: const SizedBox(),
+      isFilteredOnline: true,
+      showClearButton: true,
+      showSearchBox: true,
+      dropdownSearchDecoration: InputDecoration(
+        label: Text(widget.fieldContext.field.name),
+      ).applyDefaults(decorationTheme),
     );
   }
 }
