@@ -1,5 +1,4 @@
 import 'package:auto_form_example/api/fake_api.dart';
-import 'package:auto_form_example/entities/book.dart';
 import 'package:flutter_auto_form/flutter_auto_form.dart';
 
 class OrderForm extends TemplateForm {
@@ -10,17 +9,25 @@ class OrderForm extends TemplateForm {
       name: 'Order\'s number',
       validators: [NotNullValidator('Please enter a valid integer')],
     ),
-    AFSearchModelField<Book>(
-      id: 'book',
-      name: 'Book',
-      validators: [NotNullValidator('Please select a book')],
-      search: searchBook,
-    ),
     AFBooleanField(
       id: 'payed',
       name: 'Is payed',
       validators: [],
       value: false,
+    ),
+    AFSelectField(
+      id: 'select',
+      name: 'Select',
+      validators: [],
+      values: [1, 2],
+      textBuilder: (e) => e.toString(),
+      value: 1,
+    ),
+    AFSearchMultipleModelsField(
+      id: 'books',
+      name: 'Books',
+      validators: [NotEmptyListValidator('Please select at least one book')],
+      search: searchBook,
     )
   ];
 }
