@@ -14,6 +14,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AFTheme(
+      data: const AFThemeData(),
       child: MaterialApp(
         title: 'Auto Form Demo',
         debugShowCheckedModeBanner: false,
@@ -50,39 +51,59 @@ class _MyHomePageState extends State<MyHomePage> {
                     child: TabBarView(
                       children: [
                         FormShowcaseTile(
-                          title: 'Login form',
-                          child: AFWidget<LoginForm>(
-                            handleErrorOnSubmit: print,
-                            formBuilder: () => LoginForm(),
-                            submitButton:
-                                (Function({bool showLoadingDialog}) submit) {
+                          title: 'Json Schema Form',
+                          child: AFWidget<JsonSchemaForm>(
+                            formBuilder: () =>
+                                JsonSchemaForm.fromJson(jsonSchema),
+                            submitButton: (Function() submit) {
                               return Padding(
                                 padding: const EdgeInsets.only(top: 32),
                                 child: ElevatedButton(
                                   child: const Text('Submit'),
-                                  onPressed: () =>
-                                      submit(showLoadingDialog: true),
+                                  onPressed: () => submit(),
                                 ),
                               );
                             },
-                            onSubmitted: (LoginForm form) async {
-                              await Future.delayed(const Duration(seconds: 2));
+                            onSubmitted: (JsonSchemaForm form) {
                               print(form.toMap());
                             },
+                          ),
+                        ),
+                        SingleChildScrollView(
+                          child: FormShowcaseTile(
+                            title: 'Login form',
+                            child: AFWidget<LoginForm>(
+                              handleErrorOnSubmit: print,
+                              formBuilder: () => LoginForm(),
+                              submitButton: (Function() submit) {
+                                return Padding(
+                                  padding: const EdgeInsets.only(top: 32),
+                                  child: ElevatedButton(
+                                    child: const Text('Submit'),
+                                    onPressed: () => submit(),
+                                  ),
+                                );
+                              },
+                              onSubmitted: (LoginForm form) async {
+                                await Future.delayed(
+                                    const Duration(seconds: 2));
+                                print(form.toMap());
+                              },
+                            ),
                           ),
                         ),
                         FormShowcaseTile(
                           title: 'Registration form',
                           child: AFWidget<RegistrationForm>(
                             formBuilder: () => RegistrationForm(),
-                            submitButton:
-                                (Function({bool showLoadingDialog}) submit) {
+                            submitButton: (Function() submit) {
                               return Padding(
                                 padding: const EdgeInsets.only(top: 32),
                                 child: ElevatedButton(
                                   child: const Text('Submit'),
-                                  onPressed: () =>
-                                      submit(showLoadingDialog: true),
+                                  onPressed: () {
+                                    submit();
+                                  },
                                 ),
                               );
                             },
@@ -91,45 +112,24 @@ class _MyHomePageState extends State<MyHomePage> {
                             },
                           ),
                         ),
-                        FormShowcaseTile(
-                          title: 'Order form',
-                          child: AFWidget<OrderForm>(
-                            formBuilder: () => OrderForm(),
-                            submitButton:
-                                (Function({bool showLoadingDialog}) submit) {
-                              return Padding(
-                                padding: const EdgeInsets.only(top: 32),
-                                child: ElevatedButton(
-                                  child: const Text('Submit'),
-                                  onPressed: () =>
-                                      submit(showLoadingDialog: true),
-                                ),
-                              );
-                            },
-                            onSubmitted: (OrderForm form) {
-                              print(form.toMap());
-                            },
-                          ),
-                        ),
-                        FormShowcaseTile(
-                          title: 'Json Schema Form',
-                          child: AFWidget<JsonSchemaForm>(
-                            formBuilder: () =>
-                                JsonSchemaForm.fromJson(jsonSchema),
-                            submitButton:
-                                (Function({bool showLoadingDialog}) submit) {
-                              return Padding(
-                                padding: const EdgeInsets.only(top: 32),
-                                child: ElevatedButton(
-                                  child: const Text('Submit'),
-                                  onPressed: () =>
-                                      submit(showLoadingDialog: true),
-                                ),
-                              );
-                            },
-                            onSubmitted: (JsonSchemaForm form) {
-                              print(form.toMap());
-                            },
+                        SingleChildScrollView(
+                          child: FormShowcaseTile(
+                            title: 'Order form',
+                            child: AFWidget<OrderForm>(
+                              formBuilder: () => OrderForm(),
+                              submitButton: (Function() submit) {
+                                return Padding(
+                                  padding: const EdgeInsets.only(top: 32),
+                                  child: ElevatedButton(
+                                    child: const Text('Submit'),
+                                    onPressed: () => submit(),
+                                  ),
+                                );
+                              },
+                              onSubmitted: (OrderForm form) {
+                                print(form.toMap());
+                              },
+                            ),
                           ),
                         ),
                       ],
@@ -159,37 +159,39 @@ class FormShowcaseTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: Container(
-        margin: const EdgeInsets.symmetric(vertical: 64, horizontal: 24),
-        padding: const EdgeInsets.all(24),
-        decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(16),
-            boxShadow: const [
-              BoxShadow(
-                blurRadius: 16,
-                color: Colors.black12,
-                offset: Offset(0, 10),
-              ),
-            ],
-            color: Colors.white),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Padding(
-              padding: const EdgeInsets.only(bottom: 16),
-              child: Text(
-                title,
-                textAlign: TextAlign.center,
-                style: const TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.blueAccent,
+    return SingleChildScrollView(
+      child: Center(
+        child: Container(
+          margin: const EdgeInsets.symmetric(vertical: 64, horizontal: 24),
+          padding: const EdgeInsets.all(24),
+          decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(16),
+              boxShadow: const [
+                BoxShadow(
+                  blurRadius: 16,
+                  color: Colors.black12,
+                  offset: Offset(0, 10),
+                ),
+              ],
+              color: Colors.white),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Padding(
+                padding: const EdgeInsets.only(bottom: 16),
+                child: Text(
+                  title,
+                  textAlign: TextAlign.center,
+                  style: const TextStyle(
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                    color: Colors.blueAccent,
+                  ),
                 ),
               ),
-            ),
-            child,
-          ],
+              child,
+            ],
+          ),
         ),
       ),
     );
