@@ -3,7 +3,8 @@ library auto_form;
 export 'src/json_schema/form.dart';
 export 'src/models/field/field.dart';
 export 'src/models/form.dart';
-export 'src/models/validator/defaults.dart';
+export 'src/models/validators/defaults.dart';
+export 'src/models/validators/validator.dart';
 export 'src/widgets/form.dart';
 export 'src/widgets/form_state.dart';
 export 'src/widgets/theme.dart';
