@@ -1,34 +1,21 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_auto_form/flutter_auto_form.dart';
+import 'package:flutter_auto_form/src/models/field/field_context.dart';
 
-/// An interface for the text field builder. See [kTextFieldWidgetBuilder] for
-/// an example of a working implementation.
-typedef TextFieldWidgetBuilder<T> = Widget Function(
-  BuildContext context, {
-  String? labelText,
-  Function(String? value)? validator,
-  required TextEditingController controller,
-  required FocusNode focusNode,
-  required bool forceError,
-  required TextInputAction action,
-  required List<String> autoFillHints,
-  required bool obscureText,
-  Function()? completeAction,
+/// The interface for the [Field.widgetBuilder] method which builds
+/// a widget from information available in the [FieldContext] instance it receives.
+///
+/// Any widget that is supposed to represent a [Field] should interface with
+/// this typedef.
+typedef FieldWidgetConstructor = Widget Function({
+  Key? key,
+  required FieldContext fieldContext,
 });
 
 /// An interface for the future loader. See [kShowFutureLoadingWidget] for
 /// an example of a working implementation.
-typedef FutureLoadingWidget<T> = FutureOr<T> Function({
+typedef FutureWrapper<T> = FutureOr<T> Function({
   required BuildContext context,
   required FutureOr<T> future,
-});
-
-/// The generic interface for field widget builder.
-typedef FieldWidgetBuilder<T> = Widget Function(
-  BuildContext context, {
-  Field? field,
-  String? nextFieldId,
-  bool? isFinal,
 });
