@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_auto_form/flutter_auto_form.dart';
 import 'package:flutter_auto_form/src/json_schema/json_schema.dart';
-import 'package:flutter_auto_form/src/models/validator/validator.dart';
+import 'package:flutter_auto_form/src/models/field/field.dart';
+import 'package:flutter_auto_form/src/models/form.dart';
+import 'package:flutter_auto_form/src/models/validators/validator.dart';
 
 class JsonSchemaForm extends TemplateForm {
   JsonSchemaForm(this.fields);
@@ -12,34 +13,63 @@ class JsonSchemaForm extends TemplateForm {
     final List<Field<Object>> fields = [];
 
     for (final JsonSchemaProperty element in schema.properties) {
-      final List<Validator<Object>> validators = [];
-
-      switch (element.type) {
-        case JsonSchemaType.string:
-          fields.add(AFTextField(
-            id: element.id,
-            name: element.name,
-            validators: validators,
-            type: AFTextFieldType.TEXT,
-          ));
-          break;
-        case JsonSchemaType.number:
-          fields.add(AFNumberField(
-            id: element.id,
-            name: element.name,
-            validators: [],
-          ));
-          break;
-        case JsonSchemaType.boolean:
-          fields.add(AFBooleanField(
-            id: element.id,
-            name: element.name,
-            validators: [],
-          ));
-          break;
-        default:
-          debugPrint(
-              '[Flutter Auto Form]: ${element.type} is not yet implemented.');
+      if (element.enumValues.isNotEmpty) {
+        fields.add(AFSelectField(
+          id: element.id,
+          name: element.name,
+          validators:
+              buildValidators<List>(element, schema.isRequired(element)),
+          values: element.enumValues,
+          textBuilder: (e) => e.toString(),
+          value: element.defaultValue ?? element.enumValues.first,
+        ));
+      } else {
+        switch (element.type) {
+          case JsonSchemaType.string:
+            fields.add(AFTextField(
+              id: element.id,
+              name: element.name,
+              validators:
+                  buildValidators<String>(element, schema.isRequired(element)),
+              value: element.defaultValue,
+              type: AFTextFieldType.TEXT,
+            ));
+            break;
+          case JsonSchemaType.number:
+            fields.add(AFNumberField(
+              id: element.id,
+              name: element.name,
+              value: element.defaultValue as num?,
+              validators:
+                  buildValidators<num>(element, schema.isRequired(element)),
+            ));
+            break;
+          case JsonSchemaType.integer:
+            fields.add(
+              AFNumberField(
+                id: element.id,
+                name: element.name,
+                value: element.defaultValue as int?,
+                validators: [
+                  ...buildValidators<int>(element, schema.isRequired(element)),
+                  IsIntegerValidator('Please enter a valid integer')
+                ],
+              ),
+            );
+            break;
+          case JsonSchemaType.boolean:
+            fields.add(AFBooleanField(
+              id: element.id,
+              name: element.name,
+              value: element.defaultValue as bool? ?? false,
+              validators:
+                  buildValidators<bool>(element, schema.isRequired(element)),
+            ));
+            break;
+          default:
+            debugPrint(
+                '[Flutter Auto Form]: ${element.type} is not yet implemented.');
+        }
       }
     }
 
@@ -48,4 +78,15 @@ class JsonSchemaForm extends TemplateForm {
 
   @override
   final List<Field<Object>> fields;
+}
+
+List<Validator<T>> buildValidators<T>(
+    JsonSchemaProperty element, bool required) {
+  final List<Validator<T>> validators = [];
+
+  if (required) {
+    validators.add(NotNullValidator('This field cannot be null'));
+  }
+
+  return validators;
 }
