@@ -1,4 +1,5 @@
-import 'package:flutter_auto_form/src/models/validator/validator.dart';
+import 'package:flutter_auto_form/src/configuration/typedef.dart';
+import 'package:flutter_auto_form/src/models/validators/validator.dart';
 
 export 'defaults.dart';
 
@@ -32,7 +33,7 @@ abstract class Field<T extends Object> {
 
   /// This method returns null if the field is valid. Otherwhise it will
   /// return the error's string specified in the validator (see [Validator]).
-  String? validate(T? value) {
+  String? validator([Object? _]) {
     for (final Validator validator in validators) {
       final String? error = validator.validate(value);
 
@@ -42,5 +43,8 @@ abstract class Field<T extends Object> {
     return null;
   }
 
+  /// Parses a value into an instance of T.
   T? parser(covariant Object? unparsedValue);
+
+  FieldWidgetConstructor get widgetBuilder;
 }
