@@ -11,41 +11,57 @@ class AFWidget<T extends TemplateForm> extends StatefulWidget {
     required this.onSubmitted,
     this.submitButton,
     this.handleErrorOnSubmit,
+    this.enableFinalAction = true,
+    this.enableSubmitFormWrapper,
   }) : super(key: key);
 
   final T Function() formBuilder;
 
   final Function(T form) onSubmitted;
 
+  final bool enableFinalAction;
+
+  final bool? enableSubmitFormWrapper;
+
   final ValueChanged<String>? handleErrorOnSubmit;
 
-  final Widget Function(Function({bool showLoadingDialog}) submit)?
-      submitButton;
+  final Widget Function(Function() submit)? submitButton;
 
   @override
   AFWidgetState<T> createState() => AFWidgetState<T>(
         formBuilder(),
         handleErrorOnSubmit: handleErrorOnSubmit,
+        enableFinalAction: enableFinalAction,
+        enableSubmitFormWrapper: enableSubmitFormWrapper,
       );
 }
 
 class AFWidgetState<T extends TemplateForm>
     extends AFFormState<AFWidget<T>, T> {
-  AFWidgetState(T model, {ValueChanged<String>? handleErrorOnSubmit})
-      : super(model: model, handleErrorOnSubmit: handleErrorOnSubmit);
+  AFWidgetState(
+    T model, {
+    ValueChanged<String>? handleErrorOnSubmit,
+    required bool enableFinalAction,
+    bool? enableSubmitFormWrapper,
+  }) : super(
+          model: model,
+          handleErrorOnSubmit: handleErrorOnSubmit,
+          enableFinalAction: enableFinalAction,
+          enableSubmitFormWrapper: enableSubmitFormWrapper,
+        );
 
   @override
   Widget build(BuildContext context) {
-    widget;
     Widget child = form();
 
     if (widget.submitButton != null) {
       child = Column(
         children: [
           child,
-          widget.submitButton!(
-            ({bool showLoadingDialog = false}) => submitForm(
-              showLoading: showLoadingDialog,
+          Padding(
+            padding: const EdgeInsets.only(top: 16),
+            child: widget.submitButton!(
+              () => submitForm(),
             ),
           )
         ],
